refactor(racing): view와 domain의 분리
