+++ src/main/java/racing/domain/WinnerMaker.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import java.util.List;
