+++ src/main/java/racing/domain/strategy/DrivingStopStrategy.java
@@ -1,4 +1,4 @@
-package racing.strategy;
+package racing.domain.strategy;
 
 public class DrivingStopStrategy implements DrivingStrategy {
     @Override
