+++ src/test/java/racing/CarTest.java
@@ -2,8 +2,9 @@ package racing;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racing.strategy.DrivingMoveStrategy;
-import racing.strategy.DrivingStopStrategy;
+import racing.domain.Car;
+import racing.domain.strategy.DrivingMoveStrategy;
+import racing.domain.strategy.DrivingStopStrategy;
 
 import static org.assertj.core.api.Assertions.assertThat;
